+++ genFigures.py
@@ -82,19 +82,19 @@ if __name__ == '__main__':
         overlayCartoon(fdir + 'figureS15.svg',
                        f'{cartoon_dir}/figureS05.svg', 65, 0, scalee=0.3)
 
-    if sys.argv[1] == '3B':
+    if sys.argv[1] == '4':
         # Overlay Figure 3b cartoon
-        overlayCartoon(fdir + 'figure3B.svg',
+        overlayCartoon(fdir + 'figure4.svg',
                        f'{cartoon_dir}/figure3B.svg', 50, 40, scalee=0.24)
 
-    if sys.argv[1] == '4':
+    if sys.argv[1] == '5':
         # Overlay Figure 4 cartoon
-        overlayCartoon(fdir + 'figure4.svg',
+        overlayCartoon(fdir + 'figure5.svg',
                        f'{cartoon_dir}/figureS06.svg', 10, 5, scalee=0.27)
 
-    if sys.argv[1] == '6':
+    if sys.argv[1] == '7':
         # Overlay Figure 6 cartoon
-        overlayCartoon(fdir + 'figure6.svg',
+        overlayCartoon(fdir + 'figure7.svg',
                        f'{cartoon_dir}/lineage_notcen.svg', 0, 0, scalee=0.5)
-        overlayCartoon(fdir + 'figure6.svg',
+        overlayCartoon(fdir + 'figure7.svg',
                        f'{cartoon_dir}/lineage_cen.svg', 230, 0, scalee=0.5)
