+++ manuscript/03_results.md
@@ -12,19 +12,19 @@ Any observable heritable phenotype with a definable distribution could be used a
 ### Increasing depth and breadth of the lineage increases the accuracy of state assignment and parameter estimation.
 Here we show the performance of the model by increasing the cell numbers. The figure shows the state assignment accuracy and parameter estimations.
 
-![**A figure to show how many cells we would need to obtain a reasonable performance.](./output/figure3A.svg){#fig:3A}
+![**A figure to show how many cells we would need to obtain a reasonable performance.](./output/figure3.svg){#fig:3}
 
 
 ### tHMM performs best when the difference between cell states is significant.
 Employing Wasserestein distance, we show the more different the states, the higher the accuracy of state assignments.
 
-![**Wasserestein distance shows in a 2-state population, the accuracy increases if the states are farther to each oteher.](./output/figure3B.svg){#fig:3B}
+![**Wasserestein distance shows in a 2-state population, the accuracy increases if the states are farther to each oteher.](./output/figure4.svg){#fig:4}
 
 
 ### tHMM model can pick up even rare phenotypes within the lineage
 This parts explains how the model performs in the case of an under-represented population of cells.
 
-![**State assignment accuracy when we have different proportions of each state, showing the model performs well even when there is an under- (or over-) represented population.](./output/figure4.svg){#fig:4}
+![**State assignment accuracy when we have different proportions of each state, showing the model performs well even when there is an under- (or over-) represented population.](./output/figure5.svg){#fig:5}
 
 ### AIC metric shows the model determines the right number of states in a population.
 In this part we explain the AIC figures.
@@ -32,4 +32,4 @@ In this part we explain the AIC figures.
 ### Model performs reasonable even in the presence of different censoring types in the data.
 In this part we compare model performance in case where we have a full binary tree, and where the data in censored due to cell death and experiment end time. The upper row shows the lineage tree over time for full tree and censored data, the lower row represents state assignement accuracy for both cases.
 
-![**Censored versus uncensored data. Model performes well even if we have censored data.](./output/figure6.svg){#fig:6}
+![**Censored versus uncensored data. Model performes well even if we have censored data.](./output/figure7.svg){#fig:7}
